requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
isort==5.4.2
jws==0.1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==4.1.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
pylint==2.6.0
Pyrebase4==4.4.0
python-jwt==2.0.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
